requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.5.0
Django==1.11.11
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-pipeline==1.6.14
django-registration==2.4.1
django-registration-redux==2.2
django-suit==0.2.26
django-webpack-loader==0.6.0
djangorestframework==3.7.7
futures==3.1.1
gunicorn==19.7.1
mysqlclient==1.3.12
pytz==2018.3
whitenoise==3.3.1
psycopg2==2.7.2
